[feat] Observer created to take the data from the form
An observer was created to take the data from the forms (page 1 and page 2).
Next steps get the values from the form Dropdowns and add them to the object that store the form data.
Once the JSON gets all the data from the form, make it a string and use it for RescueService.addRescue()